pyproject.toml
[tool.poetry]
name = "sergeychartsocr"
version = "0.1.0"
description = ""
authors = ["blowyourbra1n <blowyourbra1n@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.2.0"
pytesseract = "^0.3.10"
easyocr = "^1.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
